requirements.txt
APScheduler==3.5.3
Click==7.0
ConfigArgParse==0.13.0
Flask-Cors==3.0.7
Flask==1.0.2
Jinja2==2.10.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.1.0
PyJWT==1.7.1
PyYAML==5.1.1
SQLAlchemy==1.3.10
Sanic-Cors==0.9.9.post1
Sanic-Plugins-Framework==0.8.2
Werkzeug==0.15.2
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
astor==0.8.0
async-generator==1.10
async-timeout==3.0.1
attrs==19.1.0
boto3==1.9.144
botocore==1.12.144
bz2file==0.98
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cloudpickle==0.6.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
dill==0.2.9
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
dopamine-rl==2.0.5
fbmessenger==5.3.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
gin-config==0.2.1
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.6.3
google-pasta==0.1.7
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.21.1
gunicorn==19.9.0
gym==0.15.3
h11==0.8.1
h2==3.1.1
h5py==2.9.0
hpack==3.0.0
httpcore==0.3.0
httplib2==0.14.0
httptools==0.0.13
humanfriendly==4.18
hyperframe==5.2.0
idna-ssl==1.1.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
jmespath==0.9.4
jsonpickle==1.0
jsonschema==2.6.0
kafka-python==1.4.7
kfac==0.2.0
kiwisolver==1.1.0
matplotlib==2.2.4
mattermostwrapper==2.1
mesh-tensorflow==0.1.4
more-itertools==7.2.0
mpmath==1.1.0
multidict==4.5.2
networkx==2.3
numpy==1.16.3
oauth2client==4.1.3
opencv-python==4.1.1.26
packaging==18.0
pika==0.12.0
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.8.0
pyasn1-modules==0.2.7
pyasn1==0.4.7
pycparser==2.19
pydot==1.4.1
pyglet==1.3.2
pykwalify==1.7.0
pymongo==3.7.2
pyparsing==2.4.0
pypng==0.0.20
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-engineio==3.8.1
python-socketio==3.1.1
python-telegram-bot==11.1.0
pytz==2018.9
questionary==1.0.1
rasa-sdk==1.4.0
rasa==1.4.1
redis==2.10.6
requests-async==0.5.0
requests-toolbelt==0.9.1
requests==2.21.0
rfc3986==1.3.2
rocketchat-API==0.6.25
rsa==4.0
ruamel.yaml==0.15.94
s3transfer==0.2.0
sanic-jwt==1.3.2
sanic==19.9.0
scikit-learn==0.20.2
scipy==1.3.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.0
sympy==1.4
tabulate==0.8.3
tensor2tensor==1.14.1
tensorboard==1.14.0
tensorflow-datasets==1.3.0
tensorflow-estimator==1.14.0
tensorflow-gan==2.0.0
tensorflow-hub==0.6.0
tensorflow-metadata==0.15.0
tensorflow-probability==0.7.0
tensorflow==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.31.1
twilio==6.23.0
typing-extensions==3.7.2
tzlocal==1.5.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.3
uvloop==0.13.0
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.56.0
websockets==8.0.2
wrapt==1.11.2
yarl==1.3.0
zipp==0.6.0
fasttext==0.9.1